refactor: consolidate login and registration logic into a single AuthPage component with improved styling
